refactor: hierarchy
