+++ src/models/hierarchy/impl.ts
@@ -2,9 +2,9 @@ import { TFile } from "obsidian";
 import { HierarchyNode, HierarchyNodeProps } from "./node";
 
 export type HierarchyImpl = {
-	getDataByKey: (key: string) => TFile | undefined;
+	getFileByPath: (key: string) => TFile | undefined;
 	getChildren: (data: TFile) => Promise<TFile[]>;
 	getParents: (data: TFile) => Promise<TFile[]>;
-	getKey: (data: TFile) => string;
+	getPath: (data: TFile) => string;
 	createNode: (props: HierarchyNodeProps) => HierarchyNode;
 };
