+++ src/utils/hierarchy.ts
@@ -1,171 +1,29 @@
-import { App, TFile } from "obsidian";
+import { TFile } from "obsidian";
 import {
+	checkHasPropByPosition,
 	extractMdLinkPath,
-	getFileBacklinks,
-	getFileByPath,
 	getFileName,
 	getRelativeFileByName,
 } from "./obsidian";
 import { PluginContext } from "src/context";
 import { createPromise } from "./basic";
-import { HierarchyImpl } from "src/models/hierarchy/impl";
 import { HierarchyNode } from "src/models/hierarchy/node";
 
-const checkHasPropByPosition = async (p: {
-	app: App;
-	file: TFile;
-	startOffset: number;
-	parentPropName: string;
-}): Promise<boolean> => {
-	const content = await p.app.vault.read(p.file);
+export const getActiveFileNode = (ctx: PluginContext): HierarchyNode | null => {
+	const currentFile = ctx.app.workspace.getActiveFile();
+	if (currentFile === null) return null;
 
-	let currentIndex = p.startOffset - 1;
-	const lineChars = [] as string[];
-
-	while (currentIndex >= 0) {
-		const char = content[currentIndex];
-		if (char === undefined || char === "\n") break;
-		lineChars.push(char);
-		currentIndex -= 1;
-	}
-
-	lineChars.reverse();
-	const line = lineChars.join("");
-
-	// FIXME false positive matches are possible
-	return Boolean(
-		line.match(RegExp(`(^\\s*-?\\s*|\\[)${p.parentPropName}::\\s*`)),
-	);
-};
-
-export const getParentFiles = async (p: {
-	app: App;
-	file: TFile;
-	parentPropName: string;
-}): Promise<TFile[]> => {
-	const result: TFile[] = [];
-
-	const metadata = p.app.metadataCache.getFileCache(p.file);
-
-	const frontMatterLinks = metadata?.frontmatterLinks;
-
-	if (frontMatterLinks) {
-		for (const item of frontMatterLinks) {
-			if (item.key.split(".")[0] === p.parentPropName) {
-				const linkedFile = getRelativeFileByName(
-					p.app,
-					item.link,
-					p.file.path,
-				);
-				if (linkedFile) {
-					result.push(linkedFile);
-				}
-			}
-		}
-	}
-
-	const links = metadata?.links;
-
-	if (links) {
-		for (const item of links) {
-			if (
-				await checkHasPropByPosition({
-					app: p.app,
-					parentPropName: p.parentPropName,
-					file: p.file,
-					startOffset: item.position.start.offset,
-				})
-			) {
-				const linkedFile = getRelativeFileByName(
-					p.app,
-					item.link,
-					p.file.path,
-				);
-				if (linkedFile) {
-					result.push(linkedFile);
-				}
-			}
-		}
-	}
-
-	return result;
-};
-
-export const getChildFiles = async (p: {
-	app: App;
-	file: TFile;
-	parentPropName: string;
-}): Promise<TFile[]> => {
-	const childFiles: TFile[] = [];
-
-	const backlinks = getFileBacklinks(p.app, p.file);
-
-	for (const [childPath, childLinks] of Object.entries(backlinks)) {
-		for (const childLink of childLinks) {
-			// if frontmatter
-			if ("key" in childLink && childLink.key) {
-				const [childLinkKey] = childLink.key.split(".");
-
-				if (childLinkKey === p.parentPropName) {
-					const childFile = getRelativeFileByName(
-						p.app,
-						childPath,
-						p.file.path,
-					);
-					if (childFile) {
-						childFiles.push(childFile);
-					}
-				}
-				// if inline
-			} else if ("position" in childLink) {
-				const childFile = getRelativeFileByName(
-					p.app,
-					childPath,
-					p.file.path,
-				);
-				if (
-					childFile &&
-					(await checkHasPropByPosition({
-						app: p.app,
-						file: childFile,
-						startOffset: childLink.position.start.offset,
-						parentPropName: p.parentPropName,
-					}))
-				) {
-					childFiles.push(childFile);
-				}
-			}
-		}
-	}
-
-	return childFiles;
-};
-
-export const checkHasParent = (
-	ctx: PluginContext,
-	file: TFile,
-	linkFile: TFile,
-): boolean => {
-	const fullLinkFile = getRelativeFileByName(
-		ctx.app,
-		linkFile.path,
-		file.path,
-	)?.path;
-	if (!fullLinkFile) return false;
-
-	return ctx.hierarchy
-		.getNode(file.path)
-		.hasRelativePath("parent", fullLinkFile);
+	return ctx.hierarchy.getNode(currentFile.path);
 };
 
 export const checkActiveFileHasParent = (
 	ctx: PluginContext,
-	linkFile: TFile,
-) => {
-	const currentFile = ctx.app.workspace.getActiveFile();
-	if (!currentFile) return false;
+	node: HierarchyNode,
+): boolean => {
+	const activeNode = getActiveFileNode(ctx);
+	if (!activeNode) return false;
 
-	return checkHasParent(ctx, currentFile, linkFile);
+	return activeNode.hasRelative("parent", node);
 };
 
 const getNormalizedFrontmatterArray = (
@@ -186,21 +44,21 @@ const getNormalizedFrontmatterArray = (
 export const addParentLink = async (
 	ctx: PluginContext,
 	p: {
-		file: TFile;
-		linkedFile: TFile;
+		node: HierarchyNode;
+		linkedNode: HierarchyNode;
 	},
 ): Promise<void> => {
 	const finished = createPromise<void>();
 
-	const checkLinked = () => checkHasParent(ctx, p.file, p.linkedFile);
+	const checkLinked = () => p.node.hasRelative("parent", p.linkedNode);
 
 	// Do not use registerEvent because the subscription happens automatically below.
 	const eventRef = ctx.app.metadataCache.on("resolved", () => {
 		if (checkLinked()) finished.resolve();
 	});
 
-	ctx.app.fileManager.processFrontMatter(p.file, (frontMatter) => {
-		const newValue = `[[${p.linkedFile.basename}]]`;
+	ctx.app.fileManager.processFrontMatter(p.node.data, (frontMatter) => {
+		const newValue = `[[${p.linkedNode.data.basename}]]`;
 
 		if (checkLinked()) {
 			finished.resolve();
@@ -221,11 +79,11 @@ export const addParentLink = async (
 export const removeParentLink = async (
 	ctx: PluginContext,
 	p: {
-		file: TFile;
-		linkedFile: TFile;
+		node: HierarchyNode;
+		linkedNode: HierarchyNode;
 	},
 ): Promise<void> => {
-	const links = ctx.app.metadataCache.getFileCache(p.file)?.links ?? [];
+	const links = ctx.app.metadataCache.getFileCache(p.node.data)?.links ?? [];
 
 	// Deleting inline links
 	for (const item of links) {
@@ -233,11 +91,11 @@ export const removeParentLink = async (
 			await checkHasPropByPosition({
 				app: ctx.app,
 				parentPropName: ctx.settings.get("parentPropName"),
-				file: p.file,
+				file: p.node.data,
 				startOffset: item.position.start.offset,
 			})
 		) {
-			const content = await ctx.app.vault.read(p.file);
+			const content = await ctx.app.vault.read(p.node.data);
 			let endOffset = item.position.end.offset;
 			while (endOffset < content.length) {
 				endOffset += 1;
@@ -256,20 +114,20 @@ export const removeParentLink = async (
 				content.slice(endOffset);
 
 			// Editor API is not needed here.
-			ctx.app.vault.modify(p.file, newContent);
+			ctx.app.vault.modify(p.node.data, newContent);
 		}
 	}
 
 	const finished = createPromise<void>();
 
-	const checkLinked = () => checkHasParent(ctx, p.file, p.linkedFile);
+	const checkLinked = () => p.node.hasRelative("parent", p.linkedNode);
 
 	// Do not use registerEvent because the subscription happens automatically below.
 	const eventRef = ctx.app.metadataCache.on("resolved", () => {
 		if (!checkLinked()) finished.resolve();
 	});
 
-	ctx.app.fileManager.processFrontMatter(p.file, (frontMatter) => {
+	ctx.app.fileManager.processFrontMatter(p.node.data, (frontMatter) => {
 		if (!checkLinked()) {
 			finished.resolve();
 			return;
@@ -287,11 +145,11 @@ export const removeParentLink = async (
 			const fullPath = getRelativeFileByName(
 				ctx.app,
 				path,
-				p.file.path,
+				p.node.data.path,
 			)?.path;
 			if (!fullPath) return false;
 
-			return fullPath === p.linkedFile.path;
+			return fullPath === p.linkedNode.data.path;
 		});
 
 		if (index !== -1) {
@@ -320,26 +178,6 @@ export const updateRootFile = (
 		}
 	}
 };
-
-export const createFileHierarchyImpl = (ctx: PluginContext): HierarchyImpl => ({
-	getChildren: async (file) =>
-		getChildFiles({
-			app: ctx.app,
-			file,
-			parentPropName: ctx.settings.get("parentPropName"),
-		}),
-	getParents: async (file) =>
-		getParentFiles({
-			app: ctx.app,
-			file,
-			parentPropName: ctx.settings.get("parentPropName"),
-		}),
-	getDataByKey: (key) => getFileByPath(ctx.app, key),
-	getKey: (file) => file.path,
-	createNode: (props) => {
-		return new HierarchyNode(props);
-	},
-});
 
 const getSortItemsFn = (
 	ctx: PluginContext,
